requirements.txt
python-fasthtml

aiohttp[speedups]

sentry-sdk

ruff
ipdb
